feat: Support dynamic reads of customDevices.json
If you use a customDevices.json this will now be read on every node
discovery instead of being read once when you start the server.
This makes it easier to build out z-wave networks without being forced
into network scans due to restarts

Dynamic parsing of customDevices.js is not suported but it is still read
at startup to continue supporting it. A good approach is probabl to
generate the json file and not use a js file