Fix undefined set on switches.
During initial discovery, while the setTopic value is undefined there is
an issue which arises with an nill value for command_topic on hass
discovery. This issue comes due to the "split" node compared to previous
driver, which combined both.
By checkinf if both values setTopic and getTopic are set, we make sure
to discover a value once is really complete.