fix(hass): Dimmers turn on to previous brightness
This fixes a side-effect from the fixes in zwave-js/zwavejs2mqtt/pull/290 that caused using the toggle to turn a dimmer on to make the light come on at full brightness instead of restoring the previous level. While there might be an argument for that as intentional behavior, this change also brings the behavior over MQTT in line with the behavior when using the direct ZwaveJS integration (where Zwave dimmers return to their previous brightness when the toggle is used).

The only caveat I can find in my testing is a very slight but noticeable delay on the brightness slider updating, because HA isn't sending a brightness so it doesn't preemptively move the slider.