feat(ui): Support system preference color scheme
We migrate from only supporting an explicit dark or light color scheme
to also supporting whatever the current system preference is for the
color scheme.

This allows folks to change their color scheme at the operating system
level (whether they do it automatically or manually) and have that
setting be reflected in Z-Wave JS UI. This also means that different
device preferences should follow what the device has configured (e.g. a
desktop might always be in dark mode, while a mobile device changes
based on the time of day).