chore(ui): Fix `ui.colorScheme` equality check
We had a small typo and were using the non-strict equality check for
`ui.colorScheme`. Non-strict equality checking is generally a bug
waiting to happen. It would be fine right now because we're working with
strings, but it could easily cause an issue if `ui.colorScheme` changed
its type.

We fix this by using the strict equality check.